+++ solver/src/Clause.cpp
@@ -4,11 +4,11 @@
 
 #include "../include/Clause.h"
 
-void Clause::addVariable(long varId, bool polarity) {
+void Clause::addVariable(size_t varId, bool polarity) {
     variables.emplace(varId, polarity);
 }
 
-bool Clause::removeVariable(long varId) {
+bool Clause::removeVariable(size_t varId) {
     return variables.erase(varId) == 1;
 }
 
@@ -16,10 +16,10 @@ bool Clause::isEmpty() const {
     return variables.empty();
 }
 
-std::unordered_map<long, bool>::const_iterator Clause::begin() {
+std::unordered_map<size_t, bool>::const_iterator Clause::begin() {
     return variables.begin();
 }
 
-std::unordered_map<long, bool>::const_iterator Clause::end() {
+std::unordered_map<size_t, bool>::const_iterator Clause::end() {
     return variables.end();
 }
